requirements.txt
Pillow
devtools
html_meta_data_parse
magentic
requests
playwright
pydantic
python-dotenv
slugify
unstructured
